# perf: limit rows to 1 for `get_value` and `exists`