Improve `db.exists` and `db.get_value` when used with filters from `O(N)` to `O(1)` :stonks:

Before: 

```ipython
%timeit frappe.db.exists("Stock Ledger Entry", {"item_code": "_Test Item"})
4 ms ± 1.06 ms per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

This is with JUST 200 matching rows in the database, in reality, SLEs often have millions of rows and all are getting serialized to python objects just to get discarded immediately after. 😄 


After:
```ipython
%timeit frappe.db.exists("Stock Ledger Entry", {"item_code": "_Test Item"})
143 µs ± 349 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```


ERPNext CI run: https://github.com/frappe/erpnext/actions/runs/2012032706